services/notification-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
aio-pika>=9.4.0
websockets>=12.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
redis[hiredis]>=5.0.0
